Enable -Wshadow, -Wconversion, -Wsign-conversion and remove -Wno-unused-parameter
- Remove global -Wno-unused-parameter (GCC/Clang) and /wd4100 (MSVC) suppressions
- Add -Wshadow, -Wconversion, -Wsign-conversion for GCC/Clang
- Add /w14456, /w14457, /w14244, /w14245 for MSVC
- Fix ~570 warnings across 49 files: shadow (5), unused-parameter (20),
  unused-variable (5), conversion/sign-conversion (~530), sequence-point UB (1)
- Patterns: static_cast<> for narrowing/sign conversions, [[maybe_unused]] for
  interface params, variable renames for shadow conflicts, dead code removal

+++ src/admin/gm_commands.cpp
@@ -1037,11 +1037,11 @@ void register_gm_commands(admin_system& admin, const gm_command_context& ctx)
                                        return command_result::error("Usage: /settime <hour> [minute]");
                                    }
 
-                                   int hour = cmd_ctx.args[0].int_value;
+                                   int hour = static_cast<int>(cmd_ctx.args[0].int_value);
                                    int minute = 0;
                                    if (cmd_ctx.args.size() > 1)
                                    {
-                                       minute = cmd_ctx.args[1].int_value;
+                                       minute = static_cast<int>(cmd_ctx.args[1].int_value);
                                    }
 
                                    if (hour < 0 || hour > 23)
